Create Accessing Objects Properties with Variables
Another use of bracket notation on objects is to use a variable to access a property. This can be very useful for iterating through lists of the object properties or for doing the lookup.

Note that we do not use quotes around the variable name when using it to access the property because we are using the value of the variable, not the name

Instructions

Use the playerNumber variable to lookup player 16 in testObj using bracket notation.