Create Using Objects for Lookups
Objects can be thought of as a key/value storage, like a dictionary. If you have tabular data, you can use an object to "lookup" values rather than a switch statement or an if/else chain. This is most useful when you know that your input data is limited to a certain range

Instructions

Convert the switch statement into a lookup table called lookup. Use it to lookup val and assign the associated string to the result variable.